Refactor data to Uint32Array packed RGBA, redesign cities as multi-layer noise
- Data refactor: SceneObject.data is now Uint32Array with packed RGBA per pixel.
  Colors are baked into the data — no more passing color/cityColor params.
  rebuildImageData just copies data to ImageData buffer. packColor() helper added.
  All callers updated: planet.ts, meteor.ts, game.ts, city.ts.

- City redesign: Multi-layer value noise (3 octaves) for organic mountain-like
  terrain. Tower layer (noise threshold 0.55) and spire layer (threshold 0.7)
  add skyscraper peaks. Cities grow from planet rim following exact curvature.
  Fully continuous height function — no discrete columns, no holes at any angle.

- Removed CITY_DATA_VALUE constant. City pixels identified by packed color match.
  countCityPixels takes cityPacked param.

Agent-Logs-Url: https://github.com/tfritzy/typerace.io/sessions/2ce36908-9eac-4804-a84b-1a5d7a014a2e

+++ frontend/src/games/word-defense/types.ts
@@ -5,7 +5,7 @@ export interface SceneObject {
   y: number;
   width: number;
   height: number;
-  data: Uint8Array;
+  data: Uint32Array;
   imageData: ImageData;
   bitmap: HTMLCanvasElement;
 }
